Changes property attributes type from char array to pointer
In reference to https://github.com/Mantle/MTLManagedObjectAdapter/issues/8

Also discussions in:
- https://github.com/Mantle/MTLManagedObjectAdapter/pull/21
- https://github.com/Mantle/MTLManagedObjectAdapter/pull/22

Once incorporated, this would resolve compilation issues comparing the type char array to a null pointer.

"This is a legitimate warning/error. A NULL comparison is useless because the character array is embedded in the structure allocation itself. But, ideally, this would be fixed upstream in libextobjc. -- @jspahrsummers"