Fix @keypath nullability warnings in Xcode 10
In Xcode 10, the boxing operator `@` now returns a `NSString * _Nullable` instead of `NSString *` when the parameter is `char *`. This change makes the `@keypath` and `@collectionKeypath` macros return nullable values as well. If you have CLANG_WARN_NULLABLE_TO_NONNULL_CONVERSION (-Wnullable-to-nonnull-conversion) enabled, Clang in Xcode 10 emits this warning for @keypath and @collectionKeypath expressions when assigning to a non-null `NSString *` variable/parameter:

warning: Implicit conversion from nullable pointer 'NSString * _Nullable' to non-nullable pointer type 'NSString * _Nonnull'

We can't use a pragma inside the macro defintion to silence the warning because the warning occurs through the boxing, which happens outside the macro.

This fix wraps the existing macro defintion in a `(YES).boolValue ? (NSString * _Nonnull)@(...) : (NSString * _Nonnull)nil` expression. This gives us a chance to silence the warning with an explicit cast to `_Nonnull`, and since `@(YES).boolValue` always succeeds, doesn't change the result of the macro. The additional overhead should be minimal compared to the `NSString` allocation because `@(YES)` should be a tagged pointer. (The cast (NSString * _Nonnull)nil is there to suppress a static analyzer nullability warning.)

I also had to extract the original macro definition into a separate `cStringKeypath` macro because `@collectionKeypath` depends on C strings.

I added explicit `_Nonnull` annotations to the variables in the key path tests to make sure the compiler would emit the warnings if something goes wrong.